+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Refactored `Handler`, now it is much easier to customize your own handler functionality.
 * Remove one `Box::new(FnOnce)`, thus enables better inlining for ffi function calls.
 * Fix bug: Dart struct(class) is not generated if the struct only appears in the return type #98.
+* Add `FlutterRustBridgeTimeoutMixin`. If used, a timeout exception will be thrown for ffi calls that do not return within time limit.
 
 ## 1.1.0